FTL_READ() and FTL_write() finally call _FTL_READ() and _FTL_WRITE() respectively and return each function's return value.
However, in FTL_READ() and FTL_write(), they contain _FTL_READ() and _FTL_WRITE() results in 'int' variable instead of 'int64_t'.

As a result, sometimes latencies are not accurately applied.

+++ hw/block/femu/ssd/ftl.c
@@ -154,7 +154,7 @@ void FTL_TERM(struct ssdstate *ssd)
 
 int64_t FTL_READ(struct ssdstate *ssd, int64_t sector_nb, unsigned int length)
 {
-	int ret;
+	int64_t ret;
 
 #ifdef GET_FTL_WORKLOAD
 	FILE* fp_workload = fopen("./data/workload_ftl.txt","a");
@@ -182,7 +182,7 @@ int64_t FTL_READ(struct ssdstate *ssd, int64_t sector_nb, unsigned int length)
 
 int64_t FTL_WRITE(struct ssdstate *ssd, int64_t sector_nb, unsigned int length)
 {
-	int ret;
+	int64_t ret;
 
 #ifdef GET_FTL_WORKLOAD
 	FILE* fp_workload = fopen("./data/workload_ftl.txt","a");
